+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * removed semicolons per http://cryptocoinjs.com/about/contributing/#semicolons
 * added `PPC` and `URO` support: [#4](https://github.com/cryptocoinjs/coininfo/issues/4), [#6](https://github.com/cryptocoinjs/coininfo/pull/6)
 * added bip32 support
+* removed `terst` dev dep for `assert`
 
 0.1.0 / 2014-04-14
 ------------------